# heatmap-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: T-Map API
  description: >
    REST API for a city activity visualization platform with venue management
    and loyalty features. Includes a heatmap and QR-based loyalty.
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@citymap.local
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
...
paths:
...
  /venues:
    get:
      tags: [Venues (Public)]
      summary: Get venues in viewport
      description: Returns only venues with ACTIVE moderation status inside the requested viewport
      operationId: getVenuesInViewport
      parameters:
        - name: swLat
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 55.7801
        - name: swLng
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 49.1102
        - name: neLat
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 55.7995
        - name: neLng
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 49.1355
        - name: category
          in: query
          description: Category filter. Can be repeated.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum: [food, entertainment, shopping]
      responses:
        '200':
          description: Venues in viewport
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenuePublicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /venues/search:
    get:
      tags: [Venues (Public)]
      summary: Search venues by name
      description: Returns ACTIVE public venues whose names match the requested text, case-insensitively.
      operationId: searchVenues
      parameters:
        - name: q
          in: query
          required: true
          description: Full or partial venue name.
          schema:
            type: string
            minLength: 1
            example: "Скуратов"
      responses:
        '200':
          description: Matching venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenueSearchResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /venues/{id}:
    get:
      tags: [Venues (Public)]
      summary: Get venue page
      description: Returns only venues with ACTIVE moderation status
      operationId: getVenueById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenuePublicResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /business/venues:
    get:
      tags: [Business (Owner)]
      summary: Get my venues
      operationId: getMyVenues
      description: Requires an authenticated user with the `BUSINESS_OWNER` role
      x-required-roles: [BUSINESS_OWNER]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owner venue list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VenueOwnerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Business (Owner)]
      summary: Register a venue
      operationId: createVenue
      x-required-roles: [BUSINESS_OWNER]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VenueCreateRequest'
      responses:
        '201':
          description: Venue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueOwnerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /business/venues/{id}:
    get:
      tags: [Business (Owner)]
      summary: Get my venue by id
      operationId: getMyVenueById
      x-required-roles: [BUSINESS_OWNER]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Owner venue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueOwnerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Business (Owner)]
      summary: Update venue details
      operationId: updateVenue
      x-required-roles: [BUSINESS_OWNER]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VenueUpdateRequest'
      responses:
        '200':
          description: Venue updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueOwnerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Business (Owner)]
      summary: Delete a venue
      operationId: deleteVenue
      x-required-roles: [BUSINESS_OWNER]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Venue deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/venues:
    get:
      tags: [AdminVenues]
      summary: Get venue moderation queue
      operationId: getAdminVenues
      x-required-roles: [ADMIN]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VenueModerationStatus'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Venue moderation queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVenueModerationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/venues/{id}:
    get:
      tags: [AdminVenues]
      summary: Get venue moderation details
      operationId: getAdminVenueById
      x-required-roles: [ADMIN]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venue moderation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVenueModerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/venues/{id}/verify:
    patch:
      tags: [AdminVenues]
      summary: Verify venue
      operationId: verifyAdminVenue
      x-required-roles: [ADMIN]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Venue verified and activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVenueModerationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/venues/{id}/reject:
    patch:
      tags: [AdminVenues]
      summary: Reject venue
      operationId: rejectAdminVenue
      x-required-roles: [ADMIN]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminModerationDecision'
      responses:
        '200':
          description: Venue rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminVenueModerationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable parameters ─────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ── Schemas ─────────────────────────────────
  schemas:
...
    VenueModerationStatus:
      type: string
      enum: [PENDING, ACTIVE, REJECTED]

    # ── Map / Heatmap DTOs ────────────────────
...
    VenueCreateRequest:
      type: object
      required: [name, address, lat, lng, category]
      properties:
        name:
          type: string
          example: "Скуратов Кофе"
        address:
          type: string
          example: "ул. Баумана, 9, Казань"
        lat:
          type: number
          format: double
          example: 55.7915
        lng:
          type: number
          format: double
          example: 49.1118
        category:
          type: string
          enum: [food, entertainment, shopping]
          example: "food"
        description:
          type: string
          example: "Лучший кофе в центре Казани. Зерно свежей обжарки."
        photoUrl:
          type: string
          format: uri
          example: "https://s3.local/images/skuratov_kzn.jpg"
        dishOfDay:
          type: string
          example: "Нитро кофе + Круассан"
        music:
          type: string
          example: "Indie Pop, Lo-Fi"
    VenueUpdateRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        description:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        category:
          type: string
          enum: [food, entertainment, shopping]
        photoUrl:
          type: string
          format: uri
        dishOfDay:
          type: string
        music:
          type: string
    VenuePublicResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "5085d345-3129-4e4b-972d-8bcf9a2ea9c9"
        name:
          type: string
          example: "Скуратов Кофе"
        address:
          type: string
          example: "ул. Баумана, 9, Казань"
        lat:
          type: number
          format: double
          example: 55.7915
        lng:
          type: number
          format: double
          example: 49.1118
        description:
          type: string
          example: "Лучший кофе в центре Казани"
        category:
          type: string
          enum: [food, entertainment, shopping]
          example: "food"
        photoUrl:
          type: string
          format: uri
        dishOfDay:
          type: string
          example: "Нитро кофе + Круассан"
        music:
          type: string
          example: "Indie Pop"
        peopleNow:
          type: integer
          description: Approximate number of people currently at the venue based on tx
          example: 18
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        promotions:
          type: array
          items:
            $ref: '#/components/schemas/VenuePromoResponse'
    VenueSearchResultResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "5085d345-3129-4e4b-972d-8bcf9a2ea9c9"
        name:
          type: string
          example: "Скуратов Кофе"
        address:
          type: string
          example: "ул. Баумана, 9, Казань"
        lat:
          type: number
          format: double
          example: 55.7915
        lng:
          type: number
          format: double
          example: 49.1118
        category:
          type: string
          enum: [food, entertainment, shopping]
          example: "food"
        photoUrl:
          type: string
          format: uri
    VenueOwnerResponse:
      allOf:
        - $ref: '#/components/schemas/VenuePublicResponse'
        - type: object
          properties:
            ownerId:
              type: string
              format: uuid
            h3Res9:
              type: string
              description: H3 resolution-9 index as string
              example: "89115b22b0bffff"
            moderationStatus:
              $ref: '#/components/schemas/VenueModerationStatus'
            rejectReason:
              type: string
              description: Reason for rejection (filled only if REJECTED)
              example: "Адрес не совпадает с координатами"

    # ── Promo DTOs ────────────────────────────
    VenuePromoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        venueId:
          type: string
          format: uuid
        title:
          type: string
          example: "Счастливые часы!"
        description:
          type: string
          example: "Скидка 20% на всю выпечку после 20:00"
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AdminModerationDecision:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
          example: "Заведение нарушает правила размещения (спам в названии)"
    AdminVenueModerationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
          example: "Скуратов Кофе"
        address:
          type: string
          example: "ул. Баумана, 9"
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        h3Res9:
          type: string
          example: "89115b22b0bffff"
        category:
          type: string
          enum: [food, entertainment, shopping]
        moderationStatus:
          $ref: '#/components/schemas/VenueModerationStatus'
        rejectReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminVenueModerationPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminVenueModerationResponse'
        page:
          type: integer
        size:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "ERROR_CODE"
        message:
          type: string
          example: "Human-readable message"

  # ── Reusable Error Responses ────────────────
  responses:
    BadRequest:
      description: Request validation failed (e.g. missing fields, invalid format)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_FAILED
            message: "Field 'lat' must be a valid float number."
    Unauthorized:
      description: Token is missing, expired, or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: "Access token is missing or has expired."
    Forbidden:
      description: Insufficient permissions to perform this action (e.g. USER accessing ADMIN route)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: "You do not have the required 'BUSINESS_OWNER' role to access this resource."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: "Requested venue with ID 5085d345-3129-4e4b-972d-8bcf9a2ea9c9 was not found."
    Conflict:
      description: Data conflict (e.g. duplicate email, already used promo, activation limit reached)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: "User with email kazan_guest@example.com is already registered."
...